remove unused resources.
