+++ app/src/main/res/layout/activity_feedback.xml
@@ -37,7 +37,6 @@
             android:layout_marginTop="20dp"
             android:layout_marginEnd="15dp"
             android:background="@drawable/bg_send_off"
-            android:fontFamily="@font/boston_black"
             android:gravity="center"
             android:text="@string/send"
             android:textColor="#ffffff"
@@ -55,7 +54,6 @@
             android:layout_marginEnd="10dp"
             android:layout_marginBottom="80dp"
             android:background="#F7F7F7"
-            android:fontFamily="@font/boston_semi_bold"
             android:gravity="start|top"
             android:padding="5dp"
             android:textColor="@color/bg_black"
