--- app/src/main/res/layout/activity_media_pipe.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".activity.XCollageActivity">
-
-    <com.example.collageapp.view.SegmentView
-        android:id="@+id/image_view_show"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:scaleType="fitCenter" />
-
-    <com.google.android.material.slider.Slider
-        android:id="@+id/slider_b"
-        android:valueTo="1"
-        android:valueFrom="0"
-        android:layout_width="match_parent"
-        android:layout_height="32dp"
-        android:layout_gravity="bottom"
-        android:layout_marginStart="32dp"
-        android:layout_marginEnd="32dp"
-        android:layout_marginBottom="32dp" />
-
-    <com.google.android.material.slider.Slider
-        android:id="@+id/slider_g"
-        android:valueTo="1"
-        android:valueFrom="0"
-        android:layout_width="match_parent"
-        android:layout_height="32dp"
-        android:layout_gravity="bottom"
-        android:layout_marginStart="32dp"
-        android:layout_marginEnd="32dp"
-        android:layout_marginBottom="80dp" />
-
-    <com.google.android.material.slider.Slider
-        android:id="@+id/slider_r"
-        android:valueTo="1"
-        android:valueFrom="0"
-        android:layout_width="match_parent"
-        android:layout_height="32dp"
-        android:layout_gravity="bottom"
-        android:layout_marginStart="32dp"
-        android:layout_marginEnd="32dp"
-        android:layout_marginBottom="124dp" />
-
-    <com.example.collageapp.view.LoadingLayout
-        android:id="@+id/view_loading"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:visibility="gone" />
-
-
-    <Button
-        android:id="@+id/button_type"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="48dp"
-        android:layout_marginStart="16dp"
-        android:text="Type" />
-
-
-</FrameLayout>
