+++ app/src/main/res/layout/activity_setting.xml
@@ -97,7 +97,6 @@
         android:id="@+id/textView2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fontFamily="@font/boston_black"
         android:text="@string/save_path"
         android:textColor="#ff495052"
         android:textSize="14sp"
@@ -110,7 +109,6 @@
         android:id="@+id/text_view_path"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fontFamily="@font/boston_semi_bold"
         android:text="/storage/emulated/0/Pictures/XCollage"
         android:textColor="#ff495052"
         android:textSize="8sp"
@@ -122,7 +120,6 @@
         android:id="@+id/text_view_feedback"
         android:layout_width="0dp"
         android:layout_height="77dp"
-        android:fontFamily="@font/boston_black"
         android:gravity="center_vertical"
         android:text="@string/feedback"
         android:textColor="#ff495052"
@@ -136,7 +133,6 @@
         android:id="@+id/text_view_share"
         android:layout_width="0dp"
         android:layout_height="77dp"
-        android:fontFamily="@font/boston_black"
         android:gravity="center_vertical"
         android:text="Share"
         android:textColor="#ff495052"
@@ -150,7 +146,6 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fontFamily="@font/boston_black"
         android:text="@string/version"
         android:textColor="#ff495052"
         android:textSize="14sp"
@@ -162,7 +157,6 @@
         android:id="@+id/textView6"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fontFamily="@font/boston_heavy_it"
         android:text="1.0.0"
         android:textColor="#ff999999"
         android:textSize="8sp"
