+++ app/src/main/res/layout/activity_x_collage.xml
@@ -34,7 +34,6 @@
             android:layout_height="32dp"
             android:layout_marginTop="9dp"
             android:background="@drawable/bg_close"
-            android:fontFamily="@font/boston_black"
             android:gravity="center"
             android:text="Close"
             android:alpha="0"
