+++ app/src/main/res/layout/fragment_gallery_manager_dialog.xml
@@ -25,7 +25,6 @@
         android:layout_below="@+id/view"
         android:layout_marginStart="15dp"
         android:layout_marginTop="10dp"
-        android:fontFamily="@font/boston_black"
         android:text="@string/manager"
         android:textColor="#353535"
         android:textSize="15sp" />
@@ -48,7 +47,6 @@
         android:layout_height="wrap_content"
         android:layout_below="@+id/text_view_1"
         android:layout_marginStart="15dp"
-        android:fontFamily="@font/boston_semi_bold"
         android:text="Please select 1-120 picture  "
         android:textColor="#858585"
         android:textSize="10dp" />
@@ -59,7 +57,6 @@
         android:layout_height="wrap_content"
         android:layout_below="@+id/text_view_1"
         android:layout_toEndOf="@+id/text_view_2"
-        android:fontFamily="@font/boston_semi_bold"
         android:textColor="#858585"
         android:textSize="10dp"
         tools:text="(15)" />
