+++ app/src/main/res/layout/fragment_pro_dialog.xml
@@ -51,7 +51,6 @@
             android:layout_marginStart="13dp"
             android:layout_marginTop="23dp"
             android:drawableStart="@drawable/icon_pro_dot"
-            android:fontFamily="@font/boston_black"
             android:text="Permanently remove watermark"
             android:textColor="#FFFFFF"
             android:textSize="16sp" />
@@ -62,7 +61,6 @@
             android:layout_marginStart="13dp"
             android:layout_marginTop="10dp"
             android:drawableStart="@drawable/icon_pro_dot"
-            android:fontFamily="@font/boston_black"
             android:text="Ultra HD export"
             android:textColor="#FFFFFF"
             android:textSize="16sp" />
@@ -74,7 +72,6 @@
             android:layout_marginStart="13dp"
             android:layout_marginTop="10dp"
             android:drawableStart="@drawable/icon_pro_dot"
-            android:fontFamily="@font/boston_black"
             android:text="Support us to do better"
             android:textColor="#FFFFFF"
             android:textSize="16sp" />
@@ -98,7 +95,6 @@
             <TextView
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:fontFamily="@font/boston_heavy_it"
                 android:gravity="center"
                 android:text="$0.99"
                 android:textColor="#ff385a63"
@@ -133,7 +129,6 @@
             android:layout_marginTop="10dp"
             android:layout_marginEnd="20dp"
             android:layout_marginBottom="10dp"
-            android:fontFamily="@font/boston_black"
             android:gravity="center"
             android:text="Long press to remove watermark once"
             android:textColor="#ff747477"
