+++ app/src/main/res/layout/fragment_ratio_dialog.xml
@@ -88,7 +88,6 @@
             android:layout_marginEnd="39dp"
             android:drawableStart="@drawable/icon_ratio_dialog_width"
             android:drawablePadding="2dp"
-            android:fontFamily="@font/boston_black"
             android:text="Width"
             android:textColor="#4E5557"
             android:textSize="14sp"
@@ -100,7 +99,6 @@
             android:layout_height="wrap_content"
             android:layout_marginStart="37dp"
             android:layout_marginTop="30dp"
-            android:fontFamily="@font/boston_black"
             android:drawableStart="@drawable/icon_ratio_dialog_height"
             android:drawablePadding="2dp"
             android:text="Height"
