+++ app/src/main/res/layout/item_album_bean.xml
@@ -31,7 +31,6 @@
         android:layout_height="wrap_content"
         android:layout_marginEnd="30dp"
         android:ems="3"
-        android:fontFamily="@font/boston_heavy_it"
         android:gravity="end"
         android:maxEms="3"
         android:minEms="3"
@@ -46,7 +45,6 @@
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="15dp"
-        android:fontFamily="@font/boston_black"
         android:textColor="#ff495052"
         android:textSize="14sp"
         app:layout_constraintBottom_toBottomOf="parent"
