+++ app/src/main/res/layout/item_share.xml
@@ -25,7 +25,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="8dp"
-        android:fontFamily="@font/boston_semi_bold"
         android:textColor="#6B6B6B"
         android:textSize="9sp"
         app:layout_constraintEnd_toEndOf="parent"
