+++ app/src/main/res/layout/layout_loading.xml
@@ -24,7 +24,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="5dp"
-        android:fontFamily="@font/boston_black"
         android:text="@string/loading"
         android:textColor="#ff495052"
         android:textSize="12sp"
