+++ app/src/main/res/layout/layout_x_collage.xml
@@ -34,7 +34,6 @@
         android:id="@+id/text_view_save"
         android:layout_width="wrap_content"
         android:layout_height="0dp"
-        android:fontFamily="@font/boston_black"
         android:gravity="center"
         android:paddingStart="20dp"
         android:paddingEnd="20dp"
@@ -73,7 +72,6 @@
         android:background="@drawable/bg_pic_manager_button"
         android:drawableStart="@drawable/icon_pic_manager_button"
         android:drawablePadding="2dp"
-        android:fontFamily="@font/boston_black"
         android:gravity="center"
         android:paddingStart="10dp"
         android:paddingEnd="10dp"
